The tray icon's rectangle is read in the right types
`icon_pos` on macOS and Windows divided the icon's width — whole pixels,
a u32 — by 2.0, and `tray-icon` keeps the position in floats and the size
in integers in the same struct. It compiled everywhere I looked, because
everywhere I looked was Linux, where that function is a different one
entirely. Both jobs died on it and v0.15.0 shipped without a single
desktop binary.

The Windows target was installed on this machine the whole time:
`cargo check --target x86_64-pc-windows-gnu -p tunante` takes 29 seconds
and says exactly what the runner said. It has been run now, and it is
green.

+++ apps/tunante/src/tray.rs
@@ -495,9 +495,12 @@ mod native {
     pub fn icon_pos() -> Option<(i32, i32)> {
         LIVE.with(|l| {
             let rect = l.borrow().as_ref()?.tray.rect()?;
+            // The position is in floats and the size in whole pixels — two
+            // different types in the same struct, which is what caught this
+            // out when it only ever compiled on Linux.
             Some((
-                (rect.position.x + rect.size.width / 2.0) as i32,
-                (rect.position.y + rect.size.height / 2.0) as i32,
+                rect.position.x as i32 + (rect.size.width / 2) as i32,
+                rect.position.y as i32 + (rect.size.height / 2) as i32,
             ))
         })
     }
